# backend/src/main/java/app/belay/auth/AuthController.java
package app.belay.auth;

import app.belay.auth.dto.ChangePasswordRequest;
import app.belay.auth.dto.LoginRequest;
import app.belay.auth.dto.MeResponse;
import app.belay.auth.dto.RegisterRequest;
import app.belay.auth.dto.UpdateProfileRequest;
import app.belay.common.NotFoundException;
import app.belay.user.AppUser;
import app.belay.user.UserRepository;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.servlet.http.HttpServletRequest;
import jakarta.servlet.http.HttpServletResponse;
import jakarta.validation.Valid;
import org.springframework.http.HttpStatus;
import org.springframework.security.authentication.AuthenticationManager;
import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
import org.springframework.security.core.Authentication;
import org.springframework.security.core.annotation.AuthenticationPrincipal;
import org.springframework.security.core.context.SecurityContext;
import org.springframework.security.core.context.SecurityContextHolder;
import org.springframework.security.web.context.SecurityContextRepository;
import org.springframework.transaction.annotation.Transactional;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PatchMapping;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.ResponseStatus;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/api/auth")
@Tag(name = "auth", description = "Registration, login and session management")
public class AuthController {

    private final AuthService authService;
    private final AuthenticationManager authenticationManager;
    private final SecurityContextRepository securityContextRepository;
    private final UserRepository userRepository;
    private final AuthRateLimiter rateLimiter;
    private final app.belay.storage.StorageService storageService;

    public AuthController(
            AuthService authService,
            AuthenticationManager authenticationManager,
            SecurityContextRepository securityContextRepository,
            UserRepository userRepository,
            AuthRateLimiter rateLimiter,
            app.belay.storage.StorageService storageService) {
        this.authService = authService;
        this.authenticationManager = authenticationManager;
        this.securityContextRepository = securityContextRepository;
        this.userRepository = userRepository;
        this.rateLimiter = rateLimiter;
        this.storageService = storageService;
    }

    /** Construit la réponse profil en résolvant l'URL de l'avatar (ou null). */
    private MeResponse toMe(AppUser user) {
        return MeResponse.from(user, storageService.publicUrlOrNull(user.getAvatarObjectKey()));
    }

    @PostMapping("/register")
    @ResponseStatus(HttpStatus.CREATED)
    @Operation(summary = "Register a new account, creating or joining an organization")
    @ApiResponse(responseCode = "201", description = "Account created and session established")
    @ApiResponse(responseCode = "404", description = "joinSlug does not match any organization")
    @ApiResponse(responseCode = "409", description = "Email already registered")
    public MeResponse register(
            @Valid @RequestBody RegisterRequest body, HttpServletRequest request, HttpServletResponse response) {
        rateLimiter.recordAndCheckRegister(request.getRemoteAddr());
        AppUser user = authService.register(body);
        establishSession(body.email(), body.password(), request, response);
        return toMe(user);
    }

    @PostMapping("/login")
    @Operation(summary = "Authenticate with email and password (server session cookie)")
    @ApiResponse(responseCode = "200", description = "Session established")
    @ApiResponse(responseCode = "401", description = "Invalid credentials")
    @Transactional(readOnly = true)
    public MeResponse login(
            @Valid @RequestBody LoginRequest body, HttpServletRequest request, HttpServletResponse response) {
        rateLimiter.assertLoginAllowed(body.email());
        try {
            UserPrincipal principal = establishSession(body.email(), body.password(), request, response);
            rateLimiter.resetLogin(body.email());
            return toMe(loadUser(principal));
        } catch (org.springframework.security.core.AuthenticationException e) {
            rateLimiter.recordLoginFailure(body.email());
            throw e;
        }
    }

    @GetMapping("/csrf")
    @Operation(summary = "Prime the CSRF cookie and return the current token")
    public java.util.Map<String, String> csrf(org.springframework.security.web.csrf.CsrfToken token) {
        return java.util.Map.of("headerName", token.getHeaderName(), "token", token.getToken());
    }

    @GetMapping("/me")
    @Operation(summary = "Current authenticated user profile")
    @Transactional(readOnly = true)
    public MeResponse me(@AuthenticationPrincipal UserPrincipal principal) {
        return toMe(loadUser(principal));
    }

    @PostMapping("/change-password")
    @ResponseStatus(HttpStatus.NO_CONTENT)
    @Operation(summary = "Change the current user's password")
    @ApiResponse(responseCode = "204", description = "Password changed")
    @ApiResponse(responseCode = "400", description = "Current password is incorrect or new password too weak")
    public void changePassword(
            @AuthenticationPrincipal UserPrincipal principal, @Valid @RequestBody ChangePasswordRequest body) {
        authService.changePassword(principal.id(), body.currentPassword(), body.newPassword());
    }

    @PatchMapping("/profile")
    @Transactional
    @Operation(summary = "Update the current user's display name and/or email (confirmed by current password)")
    @ApiResponse(responseCode = "400", description = "Current password is incorrect")
    @ApiResponse(responseCode = "409", description = "Email already registered")
    public MeResponse updateProfile(
            @AuthenticationPrincipal UserPrincipal principal, @Valid @RequestBody UpdateProfileRequest body) {
        return toMe(
                authService.updateProfile(principal.id(), body.displayName(), body.email(), body.currentPassword()));
    }

    @PostMapping(value = "/avatar", consumes = org.springframework.http.MediaType.MULTIPART_FORM_DATA_VALUE)
    @Transactional
    @Operation(summary = "Upload the current user's profile photo (JPEG/PNG/WebP, content-sniffed, max 5 MB)")
    @ApiResponse(responseCode = "400", description = "The file is not a supported image")
    public MeResponse uploadAvatar(
            @AuthenticationPrincipal UserPrincipal principal,
            @org.springframework.web.bind.annotation.RequestPart("image")
                    org.springframework.web.multipart.MultipartFile image) {
        byte[] bytes;
        try {
            bytes = image.getBytes();
        } catch (java.io.IOException e) {
            throw new java.io.UncheckedIOException("Failed to read uploaded file", e);
        }
        return toMe(authService.updateAvatar(principal.id(), bytes));
    }

    private AppUser loadUser(UserPrincipal principal) {
        return userRepository.findById(principal.id()).orElseThrow(() -> new NotFoundException("User not found"));
    }

    /** Authentifie et attache le contexte à la session serveur (cookie HttpOnly). */
    private UserPrincipal establishSession(
            String email, String password, HttpServletRequest request, HttpServletResponse response) {
        Authentication authentication = authenticationManager.authenticate(
                UsernamePasswordAuthenticationToken.unauthenticated(email, password));
        // Protection contre la fixation de session : nouvel identifiant après login
        request.getSession(true);
        request.changeSessionId();
        SecurityContext context = SecurityContextHolder.createEmptyContext();
        context.setAuthentication(authentication);
        SecurityContextHolder.setContext(context);
        securityContextRepository.saveContext(context, request, response);
        return (UserPrincipal) authentication.getPrincipal();
    }
}