# openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Belay API",
    "description" : "Multi-tenant climbing club management API",
    "version" : "0.1.0"
  },
  "servers" : [ {
    "url" : "/"
  } ],
...
  "paths" : {
...
    "/api/auth/register" : {
      "post" : {
        "tags" : [ "auth" ],
        "summary" : "Register a new account, creating or joining an organization",
        "operationId" : "register",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Account created and session established",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "joinSlug does not match any organization",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "Email already registered",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login" : {
      "post" : {
        "tags" : [ "auth" ],
        "summary" : "Authenticate with email and password (server session cookie)",
        "operationId" : "login",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LoginRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Session established",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Invalid credentials",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/change-password" : {
      "post" : {
        "tags" : [ "auth" ],
        "summary" : "Change the current user's password",
        "operationId" : "changePassword",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "Password changed"
          },
          "400" : {
            "description" : "Current password is incorrect or new password too weak"
          }
        }
      }
    },
    "/api/auth/avatar" : {
      "post" : {
        "tags" : [ "auth" ],
        "summary" : "Upload the current user's profile photo (JPEG/PNG/WebP, content-sniffed, max 5 MB)",
        "operationId" : "uploadAvatar",
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "image" : {
                    "type" : "string",
                    "format" : "binary"
                  }
                },
                "required" : [ "image" ]
              }
            }
          }
        },
        "responses" : {
          "400" : {
            "description" : "The file is not a supported image",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MeResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/profile" : {
      "patch" : {
        "tags" : [ "auth" ],
        "summary" : "Update the current user's display name and/or email (confirmed by current password)",
        "operationId" : "updateProfile",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateProfileRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "400" : {
            "description" : "Current password is incorrect",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "Email already registered",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MeResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/media/{objectKey}" : {
      "get" : {
        "tags" : [ "media" ],
        "summary" : "Stream a stored image (only keys belonging to the caller's organization)",
        "operationId" : "get_1",
        "parameters" : [ {
          "name" : "objectKey",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "404" : {
            "description" : "Unknown key or key outside the caller's organization",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "string",
                  "format" : "byte"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/csrf" : {
      "get" : {
        "tags" : [ "auth" ],
        "summary" : "Prime the CSRF cookie and return the current token",
        "operationId" : "csrf",
        "parameters" : [ {
          "name" : "token",
          "in" : "query",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/CsrfToken"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "string"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "CreateOrganization" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string",
            "maxLength" : 120,
            "minLength" : 0
          },
          "climbingType" : {
            "type" : "string",
            "enum" : [ "BOULDER", "ROPES", "BOTH" ]
          }
        },
        "required" : [ "climbingType", "name" ]
      },
      "RegisterRequest" : {
        "type" : "object",
        "properties" : {
          "email" : {
            "type" : "string",
            "format" : "email",
            "maxLength" : 255,
            "minLength" : 0
          },
          "password" : {
            "type" : "string",
            "maxLength" : 72,
            "minLength" : 0
          },
          "displayName" : {
            "type" : "string",
            "maxLength" : 120,
            "minLength" : 0
          },
          "createOrganization" : {
            "$ref" : "#/components/schemas/CreateOrganization"
          },
          "joinSlug" : {
            "type" : "string",
            "description" : "Slug of the organization to join as a pending member",
            "maxLength" : 140,
            "minLength" : 0
          },
          "createMode" : {
            "type" : "boolean"
          },
          "joinMode" : {
            "type" : "boolean"
          }
        },
        "required" : [ "displayName", "email", "password" ]
      },
      "MeResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "email" : {
            "type" : "string"
          },
          "displayName" : {
            "type" : "string"
          },
          "avatarUrl" : {
            "type" : "string"
          },
          "role" : {
            "type" : "string",
            "enum" : [ "OWNER", "ADMIN", "COACH", "MEMBER" ]
          },
          "status" : {
            "type" : "string",
            "enum" : [ "PENDING", "ACTIVE", "DISABLED" ]
          },
          "organization" : {
            "$ref" : "#/components/schemas/OrganizationSummary"
          }
        }
      },
      "OrganizationSummary" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "slug" : {
            "type" : "string"
          },
          "climbingType" : {
            "type" : "string",
            "enum" : [ "BOULDER", "ROPES", "BOTH" ]
          }
        }
      },
      "LoginRequest" : {
        "type" : "object",
        "properties" : {
          "email" : {
            "type" : "string",
            "format" : "email",
            "maxLength" : 255,
            "minLength" : 0
          },
          "password" : {
            "type" : "string",
            "maxLength" : 72,
            "minLength" : 0
          }
        },
        "required" : [ "email", "password" ]
      },
      "ChangePasswordRequest" : {
        "type" : "object",
        "properties" : {
          "currentPassword" : {
            "type" : "string",
            "minLength" : 1
          },
          "newPassword" : {
            "type" : "string",
            "maxLength" : 72,
            "minLength" : 0
          }
        },
        "required" : [ "currentPassword", "newPassword" ]
      },
...
      "UpdateProfileRequest" : {
        "type" : "object",
        "properties" : {
          "displayName" : {
            "type" : "string",
            "maxLength" : 120,
            "minLength" : 0
          },
          "email" : {
            "type" : "string",
            "format" : "email",
            "maxLength" : 255,
            "minLength" : 0
          },
          "currentPassword" : {
            "type" : "string",
            "minLength" : 1
          }
        },
        "required" : [ "currentPassword" ]
      },
...
      "CsrfToken" : {
        "type" : "object",
        "properties" : {
          "parameterName" : {
            "type" : "string"
          },
          "token" : {
            "type" : "string"
          },
          "headerName" : {
            "type" : "string"
          }
        }
      }
    }
  }
}